twister: add environment class
Add environment class to discover toolchain, zephyr version and other
items.
